Algorithms refactor (#2136)
