+++ algorithms/src/main/java/com/baeldung/algorithms/automata/RtFiniteStateMachine.java
@@ -1,4 +1,4 @@
-package com.baeldung.automata;
+package com.baeldung.algorithms.automata;
 
 /**
  * Default implementation of a finite state machine.
