+++ algorithms/src/main/java/com/baeldung/algorithms/automata/RtTransition.java
@@ -1,4 +1,4 @@
-package com.baeldung.automata;
+package com.baeldung.algorithms.automata;
 
 
 /**
