+++ algorithms/src/main/java/com/baeldung/algorithms/automata/State.java
@@ -1,4 +1,4 @@
-package com.baeldung.automata;
+package com.baeldung.algorithms.automata;
 
 /**
  * State. Part of a finite state machine.
