+++ algorithms/src/main/java/com/baeldung/algorithms/mcts/montecarlo/UCT.java
@@ -7,20 +7,18 @@ import java.util.List;
 import com.baeldung.algorithms.mcts.tree.Node;
 
 public class UCT {
-    final static double C = 1.41;
 
     public static double uctValue(int totalVisit, double nodeWinScore, int nodeVisit) {
-        if (nodeVisit == 0)
+        if (nodeVisit == 0) {
             return Integer.MAX_VALUE;
-        return ((double) nodeWinScore / (double) nodeVisit) + 1.41 * Math.sqrt(Math.log(totalVisit) / (double) nodeVisit);
+        }
+        return (nodeWinScore / (double) nodeVisit) + 1.41 * Math.sqrt(Math.log(totalVisit) / (double) nodeVisit);
     }
 
-    public static Node findBestNodeWithUCT(Node node) {
+    static Node findBestNodeWithUCT(Node node) {
         int parentVisit = node.getState().getVisitCount();
-        List<Node> childNodes = node.getChildArray();
-        return Collections.max(childNodes, Comparator.comparing(c -> {
-            double score = uctValue(parentVisit, c.getState().getWinScore(), c.getState().getVisitCount());
-            return score;
-        }));
+        return Collections.max(
+          node.getChildArray(),
+          Comparator.comparing(c -> uctValue(parentVisit, c.getState().getWinScore(), c.getState().getVisitCount())));
     }
 }
